+++ src/lib/dialog/dialog.service.ts
@@ -2,6 +2,7 @@ import {Injectable, Injector, TemplateRef} from '@angular/core';
 import {NwbDialogComponent} from './dialog.component';
 import {ComponentType, DomService} from '../shared/dom/dom.service';
 import {PortalInjector} from '../shared/portal/portal-injector';
+import {extendObject} from '../shared/util/object-extend';
 
 @Injectable()
 export class NwbDialogService {
@@ -37,7 +38,7 @@ export class NwbDialogService {
 
     const componentRef = this.domService.attachComponentPortal(NwbDialogComponent, injector);
 
-    componentRef.instance.config = config;
+    componentRef.instance.config = extendObject(new NwbDialogConfig(), config);
 
     componentRef.instance.afterClosed()
       .subscribe(() => {
@@ -49,13 +50,33 @@ export class NwbDialogService {
 }
 
 
-export interface NwbDialogConfig {
+export class NwbDialogConfig {
+
+  /** Dialog's content */
   message?: string;
+
+  /** Dialog's title */
   title: string;
-  okButtonText?: string;
-  closeButtonText?: string;
-  width?: string;
+
+  /** Text of the ok button. If none the button won't be displayed */
+  okButtonText?: string = '';
+
+  /** Text of the cancel button. If none the button won't be displayed */
+  cancelButtonText?: string = '';
+
+  /** Width of the dialog. */
+  width?: string = '';
+
+  /** Whether the dialog has a backdrop. */
+  hasBackdrop?: boolean = true;
+
+  /** Custom method to call when ok button is clicked */
+  okHandler?: Function = null;
+
+  /** Custom method to call when cancel button is clicked */
+  cancelHandler?: Function = null;
 }
+
 
 export interface ComponentType<T> {
   new (...args: any[]): T;